VBoxControl.cpp: size_t warning.
svn:sync-xref-src-repo-rev: r45357

+++ src/VBox/Additions/common/VBoxControl/VBoxControl.cpp
@@ -1,4 +1,4 @@
-/** $Id: VBoxControl.cpp 16152 2009-01-21 17:22:59Z noreply@oracle.com $ */
+/** $Id: VBoxControl.cpp 18509 2009-03-29 02:48:19Z knut.osmundsen@oracle.com $ */
 /** @file
  * VBoxControl - Guest Additions Command Line Management Interface
  */
@@ -1044,7 +1044,7 @@ static int enumGuestProperty(int argc, char *argv[])
      * arguments.
      */
     char const * const *papszPatterns = NULL;
-    size_t cPatterns = 0;
+    uint32_t cPatterns = 0;
     if (    argc > 1
         && !strcmp(argv[0], "-patterns"))
     {
